% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{get_db_conn}
\alias{get_db_conn}
\title{Retrieve the Active Database Connection}
\usage{
get_db_conn()
}
\value{
A live \code{DBIConnection} or a \code{Pool} object.
}
\description{
Returns the currently active database connection or connection pool
stored in the internal environment \code{.MINTDB_ENV}.
This function is primarily intended for internal use or for advanced
users who need direct access to the underlying connection object.
}
\details{
The returned object will be either:
\itemize{
\item a \code{DBIConnection} object, if the connection was established
using \code{\link[DBI:dbConnect]{DBI::dbConnect()}};
\item or a \code{Pool} object, if connection pooling was enabled through
\code{\link[pool:dbPool]{pool::dbPool()}}.
}

The function automatically checks that a valid connection exists
by calling \code{\link[=has_db_conn]{has_db_conn()}}. If no active connection is found,
an error is raised.
}
\examples{
\dontrun{
set_db_conn_info(driver = "sqlite", filepath = "data/test.sqlite")
set_db_conn()

con <- get_db_conn()
DBI::dbListTables(con)

set_db_disconn()
}

}
\seealso{
\code{\link[=set_db_conn]{set_db_conn()}}, \code{\link[=set_db_disconn]{set_db_disconn()}}, \code{\link[=has_db_conn]{has_db_conn()}}
}
