% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{set_db_disconn}
\alias{set_db_disconn}
\title{Disconnect Active Database Connection}
\usage{
set_db_disconn()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Safely close the active database connection or pool stored in \code{.MINTDB_ENV}.
If no connection exists, silently return.
}
