% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{has_db_conn}
\alias{has_db_conn}
\title{Check for an Active Database Connection}
\usage{
has_db_conn()
}
\value{
Logical \code{TRUE} if the connection is active; otherwise an error is thrown.
}
\description{
Verify that a database connection (or pool) exists and is still valid.
}
