% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{set_db_conn_info}
\alias{set_db_conn_info}
\title{Set Database Connection Information}
\usage{
set_db_conn_info(
  driver = c("mariadb", "mysql", "postgres", "sqlite", "odbc"),
  host = "",
  dbname = "",
  user = "",
  password = "",
  port = NULL,
  dsn = "",
  filepath = ""
)
}
\arguments{
\item{driver}{Character string specifying the DB type.
One of \code{"mariadb"}, \code{"mysql"}, \code{"postgres"}, \code{"sqlite"}, \code{"odbc"}.}

\item{host}{Database host name or IP address.}

\item{dbname}{Database name (ignored for SQLite if \code{filepath} is used).}

\item{user}{Username for authentication.}

\item{password}{Password for authentication.}

\item{port}{Numeric; port number. If missing, a default is chosen by DB type.}

\item{dsn}{For ODBC connections, a DSN string if available.}

\item{filepath}{For SQLite, path to the \code{.sqlite} file.}
}
\value{
Invisibly returns \code{TRUE} (for side effect).
}
\description{
Store database connection parameters as environment variables
for later use by \code{\link[=set_db_conn]{set_db_conn()}}. If the \code{port} argument is omitted,
a default port will be automatically assigned based on the database type:
\itemize{
\item MariaDB / MySQL: 3306
\item PostgreSQL: 5432
\item SQLite / ODBC: no port used
}
}
